base_image: Set logind check graphical to false.
On RaspberryPi images, with logind-check-graphical set to true, which
happens with the new 1.32 version of lightdm, there is no display
manager launched.  Set this check to false, so that the display manager
properly starts Xorg.

Tested on the ODROID-C2 that setting this on a system that does *not*
show this issue, also gets the display manager.